requirements.txt
absl-py==0.7.1
aiohttp==3.6.0
amqp==2.5.2
anyjson==0.3.3
asgiref==3.2.3
astor==0.8.0
async-timeout==3.0.1
asyncio==3.4.3
attrs==19.1.0
awscli==1.16.232
Babel==2.6.0
billiard==3.6.1.0
botocore==1.12.222
celery==4.4.0
certifi==2018.4.16
cffi==1.13.2
chardet==3.0.4
Click==7.0
cmake==3.14.4
colorama==0.3.9
configparser==3.5.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
cycler==0.10.0
dateparser==0.7.2
decorator==4.3.0
defusedxml==0.5.0
dicom==0.9.9.post1
Django==2.2.8
django-cors-headers==2.3.0
django-environ==0.4.5
django-filter==2.0.0
django-markdownx==2.0.28
django-multiselectfield==0.1.8
django-phonenumber-field==2.0.0
django-querysetsequence==0.8
django-render-block==0.5
django-rest-swagger==2.2.0
django-spa==0.2.0
django-templated-email==2.2.0
django-templated-mail==1.1.1
django-twilio==0.11.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
djoser==1.1.5
dnspython==1.16.0
docutils==0.15.2
et-xmlfile==1.0.1
eventlet==0.25.1
ez-setup==0.9
Flask==1.1.1
Flask-Cors==3.0.8
future==0.17.1
gast==0.2.2
get==1.0.3
gevent==1.3.4
gluoncv==0.4.0.post0
google-pasta==0.1.7
greenlet==0.4.13
grpcio==1.23.0
gunicorn==19.9.0
gwebsocket==0.9.9
h5py==2.8.0
idna==2.7
imageio==2.5.0
importlib-metadata==1.3.0
imutils==0.5.2
iso8601==0.1.12
itsdangerous==1.1.0
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.13.2
kiwisolver==1.1.0
kombu==4.6.7
lxml==4.4.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==3.0.3
MedPy==0.4.0
monotonic==1.5
more-itertools==8.0.2
multidict==4.5.2
mysql==0.0.2
mysql-connector==2.2.9
mysql-connector-python==8.0.18
mysqlclient==1.3.13
networkx==2.3
nose==1.3.7
numpy==1.16.4
oauthlib==2.1.0
openapi-codec==1.3.2
openpyxl==3.0.0
pafy==0.5.4
pandas==0.24.2
pdf417gen==0.7.0
phonenumberslite==8.9.10
Pillow==6.0.0
post==1.0.2
prompt-toolkit==1.0.14
protobuf==3.7.1
public==1.0.3
pyasn1==0.4.7
pycparser==2.19
pydicom==1.2.2
pydot==1.2.4
pyglet==1.4.4
Pygments==2.4.2
PyInquirer==1.0.3
PyJWT==1.6.4
pyparsing==2.2.2
pyserial==3.4
PySocks==1.6.8
python-dateutil==2.7.3
python-engineio==3.9.3
python-gsmmodem-new==0.12
python-socketio==4.3.1
python3-openid==3.1.0
pytube==9.5.1
pytz==2018.5
PyWavelets==1.0.3
PyYAML==5.1.2
query-string==1.0.2
redis==3.3.11
regex==2019.8.19
request==1.0.2
requests==2.22.0
requests-oauthlib==1.0.0
rsa==3.4.2
s3transfer==0.2.1
scikit-image==0.15.0
scikit-learn==0.21.1
scipy==1.3.0
simple-websocket-server==0.4.0
SimpleITK==1.2.0
simplejson==3.16.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
socketIO-client==0.7.2
SocksiPy-branch==1.1
soupsieve==1.9.1
South==1.0.2
spa==0.10.1
sqlparse==0.3.0
termcolor==1.1.0
titlecase==0.12.0
tqdm==4.34.0
twilio==6.15.2
tzlocal==2.0.0
uritemplate==3.0.0
urllib3==1.23
utc==0.0.3
vine==1.3.0
virtualenv==16.7.6
wcwidth==0.1.7
websocket==0.2.1
websocket-client==0.56.0
websocket-server==0.4
websockets==7.0
Werkzeug==0.15.5
whitenoise==3.2.2
wrapt==1.11.2
xlrd==1.2.0
yarl==1.3.0
zipp==0.6.0
